backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
bcrypt
